requirements.txt
absl-py>=2.0.0
appnope>=0.1.4
argon2-cffi>=21.3.0
argon2-cffi-bindings>=21.2.0
asttokens>=2.4.1
attrs>=24.2.0
backcall>=0.2.0
beautifulsoup4>=4.12.3
bleach>=6.1.0
certifi>=2024.8.30
cffi>=1.15.0
charset-normalizer>=3.4.0
cycler>=0.11.0
debugpy>=1.6.0
decorator>=5.1.1
defusedxml>=0.7.1
dm-tree>=0.1.8
docopt>=0.6.2
entrypoints>=0.4
executing>=2.1.0
fastjsonschema>=2.20.0
fonttools>=4.28.5
h5py>=3.10.0
idna>=3.10
importlib_metadata>=8.5.0
ipykernel>=6.13.1
ipython>=8.12.3
ipython-genutils>=0.2.0
ipywidgets>=7.7.0
jedi>=0.19.1
Jinja2>=3.1.4
jsonschema>=4.23.0
jsonschema-specifications>=2024.10.1
jupyter>=1.0.0
jupyter-console>=6.4.3
jupyter_client>=8.6.3
jupyter_core>=5.7.2
jupyterlab-widgets>=1.1.0
jupyterlab_pygments>=0.3.0
keras>=3.0.1
kiwisolver>=1.3.2
langdetect>=1.0.9
markdown-it-py>=3.0.0
MarkupSafe>=3.0.2
matplotlib>=3.5.1
matplotlib-inline>=0.1.7
mdurl>=0.1.2
mistune>=3.0.2
namex>=0.0.7
nbclient>=0.10.0
nbconvert>=7.16.4
nbformat>=5.10.4
nest-asyncio>=1.5.5
nltk>=3.8.1
notebook>=6.4.12
numpy>=1.23.2
openpyxl>=3.1.5
packaging>=24.1
pandas>=2.2.3
pandocfilters>=1.5.1
parso>=0.8.4
pexpect>=4.9.0
pickleshare>=0.7.5
Pillow>=9.1.0
pipreqs>=0.5.0
platformdirs>=4.3.6
prometheus-client>=0.14.1
prompt_toolkit>=3.0.48
psutil>=5.9.1
ptyprocess>=0.7.0
pure_eval>=0.2.3
pycparser>=2.21
Pygments>=2.18.0
pyparsing>=3.0.6
pyrsistent>=0.18.1
python-dateutil>=2.9.0.post0
pyzmq>=26.2.0
qtconsole>=5.3.1
QtPy>=2.1.0
referencing>=0.35.1
requests>=2.32.3
requests-oauthlib==1.3.1
requirements-parser==0.9.0
rich>=13.7.0
rpds-py>=0.20.0
scholarly==1.7.11
scikit-image>=0.24.0
scikit-learn>=1.5.2
scipy>=1.13.1
seaborn>=0.13.2
Send2Trash>=1.8.0
sentence-transformers>=3.2.1
six>=1.16.0
sklearn-pandas>=2.2.0
soupsieve>=2.6
stack-data>=0.6.3
terminado>=0.15.0
text-unidecode==1.3
textblob==0.17.1
tf-slim==1.1.0
tf_keras==2.17.0
thinc==8.2.5
threadpoolctl==3.5.0
tifffile==2024.9.20
timm==1.0.11
tinycss2>=1.3.0
torch>=2.5.0
tornado>=6.4.1
traitlets>=5.14.3
typing_extensions>=4.12.2
umap-learn>=0.5.6
urllib3>=2.2.3
wcwidth>=0.2.13
webencodings>=0.5.1
widgetsnbextension>=3.6.0
yarg>=0.1.9
zipp>=3.20.2
